fix validation after my merge-branch-into-the-trunk
add extra validation for sac and others



git-svn-id: http://svn.cri.ensmp.fr/svn/validation/trunk@1385 67b5017d-e62d-0410-a871-f72b1c62810f

+++ SAC/alphablending.c
@@ -1,8 +1,6 @@
-#include <stdlib.h>
-
-void alphablending(size_t n, float src1[n], float src2[n], float result[n], float alpha)
+void alphablending(unsigned int n, float src1[n], float src2[n], float result[n], float alpha)
 {
-    size_t i;
+    unsigned int i;
     for(i=0;i<n;i++)
         result[i]=alpha*src1[i]+(1-alpha)*src2[i];
 }
